T33-commit1- Thống kê, thêm thông báo ko cho xóa ảnh đại diện, thêm phần sản phẩm sắp hết hàng và sản phẩm bán chạy, comment các phần seotitle,position,codesku các bảng, chuyển một số tên sang màu đỏ

+++ BanHangOnline/BanHangOnline/Areas/Admin/Controllers/ProductImageController.cs
@@ -38,6 +38,10 @@ namespace BanHangOnline.Areas.Admin.Controllers
         public ActionResult Delete(int id) 
         {
             var item = _dbContext.ProductImages.Find(id);
+            if (item.IsDefault == true)
+            {
+                return Json(new { success = false });
+            }
             _dbContext.ProductImages.Remove(item);
             _dbContext.SaveChanges();
             return Json(new {success =true });
